Rename FramebufferDX9.cpp/h to FramebufferManagerDX9.cpp/h
