+++ CMakeLists.txt
@@ -1295,8 +1295,8 @@ set(GPU_D3D9
 	GPU/Directx9/DrawEngineDX9.h
 	GPU/Directx9/PixelShaderGeneratorDX9.cpp
 	GPU/Directx9/PixelShaderGeneratorDX9.h
-	GPU/Directx9/FramebufferDX9.cpp
-	GPU/Directx9/FramebufferDX9.h
+	GPU/Directx9/FramebufferManagerDX9.cpp
+	GPU/Directx9/FramebufferManagerDX9.h
 	GPU/Directx9/GPU_DX9.cpp
 	GPU/Directx9/GPU_DX9.h
 	GPU/Directx9/ShaderManagerDX9.cpp
