+++ GPU/Directx9/GPU_DX9.h
@@ -21,7 +21,7 @@
 #include <deque>
 
 #include "GPU/GPUCommon.h"
-#include "GPU/Directx9/FramebufferDX9.h"
+#include "GPU/Directx9/FramebufferManagerDX9.h"
 #include "GPU/Directx9/DrawEngineDX9.h"
 #include "GPU/Directx9/TextureCacheDX9.h"
 #include "GPU/Directx9/DepalettizeShaderDX9.h"
