+++ GPU/Directx9/StateMappingDX9.cpp
@@ -29,7 +29,7 @@
 #include "GPU/Directx9/GPU_DX9.h"
 #include "GPU/Directx9/ShaderManagerDX9.h"
 #include "GPU/Directx9/TextureCacheDX9.h"
-#include "GPU/Directx9/FramebufferDX9.h"
+#include "GPU/Directx9/FramebufferManagerDX9.h"
 #include "GPU/Directx9/PixelShaderGeneratorDX9.h"
 
 namespace DX9 {
