+++ GPU/Directx9/TextureCacheDX9.cpp
@@ -26,7 +26,7 @@
 #include "GPU/GPUState.h"
 #include "GPU/Directx9/PixelShaderGeneratorDX9.h"
 #include "GPU/Directx9/TextureCacheDX9.h"
-#include "GPU/Directx9/FramebufferDX9.h"
+#include "GPU/Directx9/FramebufferManagerDX9.h"
 #include "GPU/Directx9/ShaderManagerDX9.h"
 #include "GPU/Directx9/DepalettizeShaderDX9.h"
 #include "gfx/d3d9_state.h"
